splitting directories
