+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/izzanfi/phoneinformer/view/Communicable.java
@@ -1,4 +1,4 @@
-package id.ac.ui.cs.mobileprogramming.izzanfi.phoneinformer;
+package id.ac.ui.cs.mobileprogramming.izzanfi.phoneinformer.view;
 
 public interface Communicable {
     void displayDetails(String title, String description);
