+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/izzanfi/phoneinformer/view/fragments/SettingsFragment.java
@@ -1,11 +1,9 @@
-package id.ac.ui.cs.mobileprogramming.izzanfi.phoneinformer.fragments;
+package id.ac.ui.cs.mobileprogramming.izzanfi.phoneinformer.view.fragments;
 
-import android.os.Build;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -13,17 +11,17 @@ import androidx.fragment.app.Fragment;
 
 import id.ac.ui.cs.mobileprogramming.izzanfi.phoneinformer.R;
 
-public class PhoneInformationFragment extends Fragment {
+public class SettingsFragment extends Fragment {
 
-    public PhoneInformationFragment() {
+    public SettingsFragment() {
 
     }
 
-    public static PhoneInformationFragment newInstance(String title, String description) {
+    public static SettingsFragment newInstance(String title, String description) {
         Bundle args = new Bundle();
         args.putString("title", title);
         args.putString("description", description);
-        PhoneInformationFragment fragment = new PhoneInformationFragment();
+        SettingsFragment fragment = new SettingsFragment();
         fragment.setArguments(args);
 
         return fragment;
@@ -33,10 +31,7 @@ public class PhoneInformationFragment extends Fragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         super.onCreateView(inflater, container, savedInstanceState);
-        View view = inflater.inflate(R.layout.fragment_phone_information, container, false);
-        TextView content = view.findViewById(R.id.phoneInformationFragmentContentText);
-        content.setText("Phone Manufacturer: " + Build.MODEL);
-
+        View view = inflater.inflate(R.layout.fragment_settings, container, false);
         return view;
     }
 }
